test-requirements.txt
pytest==7.4.3
pytest-django==4.7.0
playwright==1.40.0
